pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "Users microservice"
authors = ["Ravil Latypov <ravillatypov12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
tortoise-orm = "^0.19.1"
asyncpg = "^0.25.0"
email-validator = "^1.2.1"
bcrypt = "^3.2.2"
aiosmtplib = "^1.1.6"
PyJWT = "^2.4.0"
orjson = "^3.7.6"
sentry-sdk = "^1.6.0"
aiofile = "^3.7.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"
Faker = "^13.14.0"
requests = "^2.28.1"
wemake-python-styleguide = "^0.16.1"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
